Added sounds and functionality to weapons
Weapons can now be purchased and used as well as make sounds when shooting and dealing damage. Fixed a few bugs such as the money counter not resetting upon death and a few visual items. The player is now allowed to deselect the weapons they are using and have a free hand.